requirements.txt
discord.py[voice]==2.5.2
PyNaCl==1.5.0
asyncmy==0.2.10
python-dotenv==1.0.1
pillow==11.1.0
gTTS==2.5.4
python-dateutil==2.9.0
psutil==7.0.0